backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
scipy==1.14.1
numpy==2.1.1
python-dotenv==1.0.1
aiosqlite==0.20.0
